requirements.txt
discord.py==2.3.2
Flask==3.0.0
flask[async]==3.0.0
Requests==2.31.0
twilio==8.10.3
